+++ generate-docs.sh
@@ -1,0 +1,5 @@
+#! /bin/bash
+
+cd ./roles/mdt_drivers
+
+aar-doc . markdown